o/assertstate: move fakeStore into assertstatetest
